@paulmillr 2021-07-21
Can you run the bench a few more times?

Get pub key is not important to us, sign and verify is